subida inicial V3 Jesus Cueto Gonzalez

+++ routes/users.js
@@ -1,5 +1,6 @@
 const router = require("express").Router();
 const { default: checkauth } = require("./middelware/checkauth");
+const Token = require("./models/token");
 const User = require("./models/users");
 
 //routes
@@ -17,11 +18,11 @@ router.get('/',checkauth.isAccessTokenValid, async (req, res)=>{
 // crear usuario, comprueba si existe y si no lo crea
 router.post('/',checkauth.isAccessTokenValid, async(req, res)=>{
     const {dni,name,last_name,mail} = req.body
-    const exist = await User.findOne({where: { mail:mail }});
+    const exist = await User.findOne({where: { dni:dni }});
     if (exist){
         return res
         .status(409)
-        .json({error: "alredy exist an accoun with the given email"});
+        .json({error: "alredy exist an accoun with the given dni"});
     }else{
         try {
             const usercreate = await User.create({ name: name,last_name:last_name, mail:mail, dni:dni});
@@ -77,19 +78,36 @@ router.get('/:user_id',checkauth.isAccessTokenValid, async (req, res)=>{
 
 
 //borrar usuario, comprobando que exista
-router.delete("/:userId",checkauth.isAccessTokenValid,async(req, res)=>{
-    const {userId} = req.params;
+router.delete("/:userId/:id",checkauth.isAccessTokenValid,async(req, res)=>{
+    const {userId,id} = req.params;
     const user = await User.findOne({where: { id:userId }});
+    const token = await Token.findOne({where: { id_usuario:userId }});
+   if(id === userId){
+    res.status(403).json("user with session open");
+   
+   }else{
     if (user){
-            try {
-                await User.destroy({ where:{id: userId }})
-                res.status(204).json({ msg: "user deleted"});   
-            } catch (error) {
-                res.status(400).json({error});
-            }      
+        try {
+            if(token){
+                await Token.destroy(
+                {
+                    where: {
+                        id_usuario: userId
+                    }
+                });
+            }
+            await User.destroy({ where:{id: userId }})
+
+            res.status(204).json({ msg: "user deleted"});   
+        } catch (error) {
+            res.status(400).json({error});
+        }      
     }else{
         res.status(400).json("user don't exist");
     }   
+
+   }
+   
 });
